Update work-around for FreeBSD p5-IO-Tty package.
FreeBSD got a new version 1.20 which is still not fixed.  Unfortunately
my earlier workaround would only do a downgrade 1.18->1.17, thinking
that the next release would surely be fixed, so we'd get a smooth
transition when the new package showed up.  But it's still broken, so we
must now downgrade 1.20->1.17.

Fortunately the real fix is coming, hopefully in the next package, based
on:

https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=276535