bsd: clear_user_and_ssh_keys script should not fail if metadata does not exist
clear_user_and_ssh_keys script does not fail when the GCP_PROJECT is
'pg-ci-images-dev' but fails when it is 'pg-ci-images'. Most possible
reason is 'pg-ci-images' project does not have at least one of the
project_level_ssh_keys and instance_level_ssh_keys. So as a solution,
script should not fail if it could not find these keys (on 404 errors).